Sankey fix:Tooltip format is not working  (T1090444) apply options.format to weight value in default templates for links and + test
